pipe: fix deadlock when piping large messages to less
We have the ability to pipe data to a QuickTerm via its stdin argument.
This function creates a virtual terminal and uses its OnStart callback
to read from stdin and write it to the created command pipe.

The OnStart callback is invoked in app.Terminal.Draw but in that case,
it waits until all data has been read and copied before returning which
prevents the main thread from actually rendering less displaying that
data on screen which blocks everything.

Keep the data copying completely async and out of the main thread to
avoid dead locking.

Reported-by: Simon Martin <simon@nasilyan.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Simon Martin <simon@nasilyan.com>