maildir: fix filename encoded UID removal when moving messages
Recently, in the process of updating the codebase to go-maildir v0.6.0
the logic to strip potential UID parts from the message filenames was
dropped. However, maildir synchronizer isync/mbsync encodes the UID
within the filename of the email and cannot recover if the UID is
preserved during a move.

Re-add the UID stripping logic to remove the "U=xxx" part from message
filenames.

Fixes: 730624ce ("mod: update go-maildir")
Link: https://lists.sr.ht/~rjarry/aerc-devel/patches/34893
Link: https://lists.sr.ht/~rjarry/aerc-discuss/%3CDACSXDU86DC3.LPLLI44C0QAK@sbinet.org%3E
Signed-off-by: Folker Schwesinger <dev@folker-schwesinger.de>
Tested-by: Sebastien Binet <s@sbinet.org>
Acked-by: Robin Jarry <robin@jarry.cc>