ui: allow message split viewer to automatically mark messages as read
Users can specify that messages should be marked as read when they're
viewed, but not when they're previewed in a message list split.

This patch makes it possible via two configuration knobs: one stating
whether the split message viewer should mark messages as read (off by
default), and one specifying the delay after which to do it.

Fixes: https://todo.sr.ht/~rjarry/aerc/126
Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>