unsubscribe: properly send unsubscribe requests
According to RFC 8058, links in the List-Unsubscribe headers should have
a POST request sent to them. This is not possible in the current setup,
as links are only opened in a browser.

Change the default behaviour to match the RFC and keep an "Open in
Browser" option for non-compliant senders.

Link: https://datatracker.ietf.org/doc/html/rfc8058
Changelog-fixed: `:unsubscribe` now follows RFC 8058.
Signed-off-by: Moritz Poldrack <git@moritz.sh>
Acked-by: Robin Jarry <robin@jarry.cc>