worker: fix race condition
There is a race condition where a callback may never be called if the
worker finishes handling a message before the callback is registered.

In practice, with Go 1.25.0, I have not seen the callback lose this
race. However, adding an explicit runtime.Gosched() before registering
the callback does cause it to consistently lose the race. Therefore,
it's probably best to patch this as whether to yield the Goroutine is
outside our control.

Signed-off-by: Marco Munizaga <git@marcopolo.io>
Reviewed-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>