export-mbox: fix occasional data loss in export result
The uids slice used inside the worker action was mutated in the main
command loop. As a message was successfully processed, its uid was
removed from the slice in a way where the last item of the original
slice was getting duplicated at the end of the underlying array.
Depending on the scheduling of the concurrent messages processing, this
was sometimes leading to some messages being replaced in the export by
the last message.

This was apparently present since the introduction of this command in
commit 845763cb1f1f8b7acdfc8e94e0a2d61ff78f6b9d.

Link: https://lists.sr.ht/~rjarry/aerc-discuss/%3CDCEVA20WLPIL.3H52WZTOTXK54@tardypad.me%3E
Changelog-fixed: Fixed occasional data loss in export result of
 `export-mbox`.
Signed-off-by: Damien Tardy-Panis <damien.dev@tardypad.me>
Acked-by: Robin Jarry <robin@jarry.cc>