msglist: make clicking messages work on initial load
I noticed that clicking on a message in the message list only opens that
message after I've navigated to another folder and came back.

The problem is that we store the number of UIDs in the list in an
attribute and screw up :-)

It's easier to simply ask the store for the number of UIDs it has when
needed, and it *also* makes clicks work on the initial message list
load.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>