maildir: add missing detection of all new messages on startup
If check-mail is not set up to run on startup (for example when the sync
is done externally), the new messages were not displayed in the dirlist
until one would switch to each of the directories.

We're now running the SyncNewMail in handleListDirectories on all
directories, the same way it's done at the end of handleCheckMail. This
registers all the new messages before the dirlist gets rendered.

Fixes: https://todo.sr.ht/~rjarry/aerc/185
Link: https://lists.sr.ht/~rjarry/aerc-discuss/%3CCSLCE7V6PBJA.THVFF075ARCT%40walnut%3E
Signed-off-by: Damien Tardy-Panis <damien.dev@tardypad.me>
Acked-by: Robin Jarry <robin@jarry.cc>