msgviewer: fix crash trying to render a missing header
I've setup aerc to use a message split and display the Message-Id
header, and have been experiencing a few crashes upon (I think) opening
new messages.

Here's an excerpt of the associated stack trace:
  [signal SIGSEGV: segmentation violation code=0x1 addr=0x0 pc=0x10967aa01]

  goroutine 1 [running]:
  git.sr.ht/~rjarry/aerc/lib/log.PanicHandler()
          git.sr.ht/~rjarry/aerc/lib/log/panic-logger.go:51 +0x6cf
  panic({0x10994e6c0?, 0x10a072b40?})
          runtime/panic.go:792 +0x132
  git.sr.ht/~rjarry/aerc/app.fmtHeader(0xc0045ab130, {0xc00011ee8c, 0xa}, {0x109952f40, 0x1b}, {0x0, 0x0}, {0x0, 0x0}, {0x0, ...})
          git.sr.ht/~rjarry/aerc/app/msgviewer.go:193 +0x381

The problem is that we happily dereference msg.RFC822Headers in
fmtHeaders, that happens to be nil. While it's not clear why that
happens (it could be linked to the fact we're working on a message we're
fetching (?)), we should definitely check RFC822Headers against nil
before dereferencing it.

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>