imap: report label updates in status bar
I noticed that changing labels does not show the "labels updated"
message in the status bar while it should. The problem is that we don't
report a "Done" in handleModifyLabels, and the callback updating the
status bar is never invoked.

This patch fixes this, along with two missing or misplaced return-s in
handleModifyLabels.

Fixes: 8f9cd5c0 ("imap: store and retrieve labels")
Fixes: 96ed6a95 ("imap: add support for :modify-labels / :tag for gmail and proton")
Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>