doc: fix example for hooks including if statements
I have tried hooks from `aerc-config` examples and noticed I keep
getting errors when they were supposed to be run. I checked the logs and
noticed that they were parsed only up to the end of if statement
(up to the ;).

Investigating further, I found that `;` is treated as a
comment in INI file format and found examples in the library used by
aerc (go-ini), showing that value needs to be wrapped in backticks to
use semicolons in it.

I have updated the config documentation to match that. I have also
tested it and it works as expected.

Link: https://github.com/go-ini/ini/blob/b2f570e5b5b844226bbefe6fb521d891f529a951/testdata/full.ini#L71
Signed-off-by: Ensar Sarajčić <dev@ensarsarajcic.com>
Acked-by: Robin Jarry <robin@jarry.cc>