help: list pages dynamically from mandb
Instead of having a hard-coded list of help topics, change the :help
command completion to list available man pages as listed by:

	man -k aerc

Append each man page short description to the completion choices.

In case mandb isn't up to date, or if running aerc without installing
it, fallback to a hard-coded list. I didn't try to walk the "doc" folder
since the working directory isn't guaranteed.

Also, in order for this to work properly, mandb must be executed after
installing the man pages. This isn't part of the makefile since it would
cause issues in downstream builds which install in staging directories.

Add a mention of mandb in the installation instructions.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Antonin Godard <antonin@godard.cc>