imap: mention cache location in error logs
Somehow one of my accounts' cache was locked, and I noticed the
following error in the logs:
  ERROR 2025/11/21 10:30:54.006464 cache.go:57: [GMail] failed opening
  cache db: resource temporarily unavailable

It's not super useful, since in particular it does not tell where it
tried to find the cache. This patch adds this location, making the error
log a bit more actionable (e.g. I know what to go `rm -rf` :-)):
  ERROR 2025/11/21 10:32:43.520762 cache.go:57: [GMail] failed opening
  cache db at /home/simon/.cache/aerc/GMail: resource temporarily
  unavailable

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>