imap: don't use list-status to discover the list of directories
The `restrict-deletes` functionality only partially worked on IMAP
except for Proton Mail, because the Junk and Trash folders are not
properly detected.

Angel Sartonev found out that they are when using LIST instead of
LIST-STATUS to discover the list of directories, which makes sense: LIST
is about directories, LIST-STATUS about their contents.

We must use LIST to discover that list, even if the IMAP provider
supports LIST-STATUS. The performance impact for correctness, if any, is
minimal (and if present, likely an improvement) since we do CheckMail
right after, hence we turn two consecutive calls to LIST-STATUS into one
call to LIST followed by one to LIST-STATUS.

Fixes: de2006bb ("config: allow controlling whether :delete is allowed from any folder")
Reported-by: Angel Sartonev <personal@asart.bg>
Reported-by: bonswouar (IRC)
Co-authored-by: Angel Sartonev <personal@asart.bg>
Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>