imap: add option to get go-imap to emit and store its own logs
I thought several times in the past that it'd be great to see how
go-imap and a given IMAP server interact, and it turns out that there's
a easy way to do it. This patch adds an option to control this.

I have validated that this logging works fine when the target path
exists and is writable to, and that we simply ignore the option
otherwise, logging something like:
  TRACE 2025/11/21 14:49:41.924753 connect.go:127: [ProtonMail] Not
  logging IMAP debug logs: open ./idontexist/imap_debug_ProtonMail.log:
  no such file or directory

Signed-off-by: Simon Martin <simon@nasilyan.com>
Acked-by: Robin Jarry <robin@jarry.cc>