ui: allow centering the message viewer and composer
On wide screens, it is common to have aerc running in a wide window. In
that case, the message viewer and the composer aligned to the left leave
the user looking at the side of the window for long periods of time,
which can be tiresome if the window sits on a side of the screen.
Introduce an option to allow the message viewer and the composer to be
centered on the window by setting a fixed width for them and padding
both sides with empty space.

Changelog-added: The editor and the message viewer can now be centered
 on screen using a fixed width.
Signed-off-by: inwit <inwit@sindominio.net>
Tested-by: Antonin Godard <antonin@godard.cc>
Acked-by: Robin Jarry <robin@jarry.cc>